move files around to reflect their use
